added data and display it to ui in home page
